@@ -6,7 +6,7 @@ authors = [
 ]
 dependencies = []
 readme = "README.rst"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 license = { text = "MIT" }
 dynamic = ["version"]
 
