@@ -9,14 +9,13 @@ setup(name='mockito',
       license='MIT',
       description='Spying framework',
       long_description=open('README.rst').read(),
-      python_requires='>=3.7',
+      python_requires='>=3.8',
       classifiers=[
           'Development Status :: 5 - Production/Stable',
           'Intended Audience :: Developers',
           'License :: OSI Approved :: MIT License',
           'Topic :: Software Development :: Testing',
           'Programming Language :: Python :: 3',
-          'Programming Language :: Python :: 3.7',
           'Programming Language :: Python :: 3.8',
           'Programming Language :: Python :: 3.9',
           'Programming Language :: Python :: 3.10',
