@@ -19,10 +19,14 @@
 # THE SOFTWARE.
 
 from __future__ import annotations
-from typing import TYPE_CHECKING
+import weakref
+from typing import TYPE_CHECKING, Callable
 
 if TYPE_CHECKING:
     from .mocking import Mock
+
+
+RegisterObserver = Callable[[object, "Mock"], None]
 
 
 class MockRegistry:
@@ -34,12 +38,52 @@ class MockRegistry:
 
     def __init__(self):
         self.mocks = IdentityMap()
+        self._register_observers: list[weakref.WeakMethod] = []
 
     def register(self, obj: object, mock: Mock) -> None:
         self.mocks[obj] = mock
 
+        alive_observers: list[weakref.WeakMethod] = []
+        for observer_ref in self._register_observers:
+            observer = observer_ref()
+            if observer is None:
+                continue
+
+            observer(obj, mock)
+            alive_observers.append(observer_ref)
+
+        self._register_observers = alive_observers
+
+    def add_register_observer(self, observer: RegisterObserver) -> None:
+        self._prune_dead_register_observers()
+        for observer_ref in self._register_observers:
+            callback = observer_ref()
+            if callback is not None and callback == observer:
+                return
+
+        self._register_observers.append(weakref.WeakMethod(observer))
+
+    def remove_register_observer(self, observer: RegisterObserver) -> None:
+        self._prune_dead_register_observers()
+
+        for i, observer_ref in enumerate(self._register_observers):
+            callback = observer_ref()
+            if callback is not None and callback == observer:
+                del self._register_observers[i]
+                break
+
+    def _prune_dead_register_observers(self) -> None:
+        self._register_observers = [
+            observer_ref
+            for observer_ref in self._register_observers
+            if observer_ref() is not None
+        ]
+
     def mock_for(self, obj: object) -> Mock | None:
         return self.mocks.get(obj, None)
+
+    def obj_for(self, mock: Mock) -> object | None:
+        return self.mocks.lookup(mock)
 
     def unstub(self, obj: object) -> None:
         try:
@@ -82,6 +126,12 @@ class IdentityMap(object):
         for k, value in self._store:
             if k is key:
                 return value
+        return default
+
+    def lookup(self, value, default=None):
+        for key, v in self._store:
+            if v is value:
+                return key
         return default
 
     def values(self):
